+++ UI/webpack.config.js
@@ -69,7 +69,11 @@ module.exports = {
         ],
       },
       {
-        test: /\.(png|jpe?g|gif|svg)$/i,
+        test: /\.svg$/i,
+        type: "asset/inline",
+      },
+      {
+        test: /\.(png|jpe?g|gif)$/i,
         type: "asset/resource",
         generator: {
           filename: "images/[name][ext][query]",